# POC: Optimizations to make data generation faster (3-7x faster)