- Draft as it builds on https://github.com/clflushopt/tpchgen-rs/pull/12
- part of https://github.com/clflushopt/tpchgen-rs/issues/6

## Why

I couldn't help myself  😎  (While testing the `dbgen` command line code I wanted to make it faster)

## What

## Performance Results

On my M3 Mac:

| Scale Factor | Main | This PR | times faster |
|--------|--------|--------|--------|
| 0.001 | 17.6s | 2.5s | 7 |
| 1 | 24.1s | 7.95 | 3 |

Note I only applied the "don't copy strings" technique to lineitem. The same thing can be done to all the other tables to get better performance


```
time cargo run --release -- -s 0.001 --output-dir=/tmp/tpchdbgen-rs
...
real	0m17.596s
```


# Background

Some quick profiling on main shows large amounts of time initializing the TextPool and string manipulaton

![Screenshot 2025-03-14 at 6 50 33 AM](https://github.com/user-attachments/assets/43022ac8-eb5f-466b-88d0-042860d55bc9)

## Changes
1. Initialize TextPool once, rather than for each table, in 15f3002179d8ca1267b9000829e1b0ded870763e
2. Avoid copying strings in LineItem (use `&str` instead), in c529cf5359b0b76a28c077abfc53d0d714b05700

Here is the profiling after these changes:

<img width="1381" alt="Screenshot 2025-03-14 at 8 22 38 AM" src="https://github.com/user-attachments/assets/2a9bc66d-f2a5-4490-89a8-d3f4923ddbbe" />

Note: I think we could reduce more of the time by formatting the date strings directly into the output (rather than into directly into buffers and then output):