@clflushopt 2025-03-15
This is looking good ! I actually had the change to switch to `Arc<TextPool>` stashed while I was finishing #16. I will clean up and finish #12 and merge this PR into it directly.