@alamb 2025-03-15
> This is looking good ! I actually had the change to switch to `Arc<TextPool>` stashed while I was finishing #16. I will clean up and finish #12 and merge this PR into it directly.

These are pretty small change and I think it will be easier to review if I pull it into a separate PR for review. ANd I an kind of procrastinating working on this (which I find fun!) before launching into more PR reviews so I'll spend a few minutes prepping