@alamb 2025-03-15
I couldn't help myself and pushed commits to port over the other tables / generators (again I will break these up as separate PRs but I wanted to see how fast I could get it)

Scale factor 1 now takes 6.68 seconds 

```
real	0m6.689s
```