@alamb 2025-03-15
BTW it takes 30 seconds to make tpch SF 100

![image](https://github.com/user-attachments/assets/de3554b7-ee7c-4d03-a630-0bae5cb82563)

```
cargo run --release -- -s 10 --output-dir=/tmp/tpchdbgen-rs-10
...
real	0m29.888s
```