@alamb 2025-03-17
> @alamb this is amazing !

BTW I have been 🚿  dreaming about this and I have some ideas on how to generate data (for a single table) using multiple threads that I am also pretty stoked about